SDI-88: ♻️  Convert app to typescript
