+++ package.json
@@ -126,7 +126,7 @@
     "prom-client": "^14.0.1",
     "querystring": "^0.2.1",
     "redis": "^4.1.1",
-    "superagent": "^7.1.6",
+    "superagent": "^8.0.0",
     "url-value-parser": "^2.1.0",
     "uuid": "^8.3.2"
   },
@@ -151,8 +151,8 @@
     "@types/superagent": "^4.1.15",
     "@types/supertest": "^2.0.12",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.30.4",
-    "@typescript-eslint/parser": "^5.30.4",
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
     "concurrently": "^7.2.2",
     "cookie-session": "^2.0.0",
     "cypress": "^10.3.0",
@@ -160,7 +160,7 @@
     "eslint": "^8.19.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-import-resolver-typescript": "^3.2.1",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-no-only-tests": "^2.6.0",
