+++ server/middleware/setUpWebSecurity.ts
@@ -1,14 +1,11 @@
 import express, { Router } from 'express'
 import helmet from 'helmet'
 
-export function setUpWebSecurity(): Router {
+export default function setUpWebSecurity(): Router {
   const router = express.Router()
 
   router.use(helmet({ contentSecurityPolicy: false }))
   router.use(helmet.referrerPolicy({ policy: 'same-origin' }))
 
   return router
-}
-export default {
-  setUpWebSecurity,
 }
